Extract StressTest to utils
